refactor: move getTimeChipText utility to a separate file for better organization
