+++ src/components/user/common/Chip.tsx
@@ -1,11 +1,6 @@
 import cn from '@/utils/cn';
-import {
-  formatTime,
-  getDday,
-  getTimeLeft,
-  isToday,
-  parseISOString,
-} from '@/utils/formatDate';
+import { parseISOString } from '@/utils/formatDate';
+import getTimeChipText from '@/utils/getTimeChipText';
 import { useEffect, useState } from 'react';
 
 const Chip = ({
@@ -36,41 +31,6 @@ const Chip = ({
   );
 };
 
-export const getTimeChipText = ({
-  open,
-  deadline,
-}: {
-  open: string;
-  deadline: string;
-}): string => {
-  const now = new Date();
-  const openTime = parseISOString(open);
-  const closeTime = parseISOString(deadline);
-
-  const timeUntilOpen = openTime.getTime() - now.getTime();
-  const timeUntilClose = closeTime.getTime() - now.getTime();
-
-  if (timeUntilOpen <= 0 && timeUntilClose <= 0) return 'CLOSED';
-
-  // 오픈 전
-  if (timeUntilOpen > 0) {
-    return isToday({ d1: openTime })
-      ? `${formatTime({ date: openTime })} OPEN`
-      : `D-${getDday(closeTime)}`;
-  }
-
-  // 오픈 후
-  const daysUntilClose = getDday(closeTime) - 1;
-  if (daysUntilClose >= 1) return 'NOW OPEN';
-
-  if (daysUntilClose === 0 && timeUntilClose > 0) {
-    const { hours, minutes, seconds } = getTimeLeft(closeTime);
-    return `${hours}:${minutes}:${seconds} LEFT`;
-  }
-
-  return 'CLOSED';
-};
-
 export const TimeChip = ({
   open,
   deadline,
@@ -90,11 +50,19 @@ export const TimeChip = ({
 
   const text = getTimeChipText({ open, deadline });
   if (text === 'CLOSED') return;
-  return <Chip theme="light-red">{text}</Chip>;
+  return (
+    <Chip rounded={rounded} theme="light-red">
+      {text}
+    </Chip>
+  );
 };
 
-export const SoldOutChip = () => {
-  return <Chip theme="grey">SOLDOUT</Chip>;
+export const SoldOutChip = ({ rounded }: { rounded?: boolean }) => {
+  return (
+    <Chip rounded={rounded} theme="grey">
+      SOLDOUT
+    </Chip>
+  );
 };
 
 export const ExtendChip = ({
